jwt rest

+++ src/main/java/org/donnex/user/model/Role.java
@@ -5,6 +5,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.OneToMany;
 import javax.persistence.Table;
 
 import org.springframework.security.core.GrantedAuthority;
@@ -13,30 +14,35 @@ import org.springframework.security.core.GrantedAuthority;
 @Table(name = "role")
 public class Role implements GrantedAuthority {
 	private static final long serialVersionUID = 1L;
-	
+
 	@Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-	@Column(name="role_id")
-	private int id;
-	@Column(name="role")
+	@GeneratedValue(strategy = GenerationType.AUTO)
+	@Column(name = "role_id")
+	private Long id;
+	
+	@Column(name = "role")
 	private String role;
 	
-	public int getId() {
+
+	public Long getId() {
 		return id;
 	}
-	public void setId(int id) {
+
+	public void setId(Long id) {
 		this.id = id;
 	}
+
 	public String getRole() {
 		return role;
 	}
+
 	public void setRole(String role) {
 		this.role = role;
 	}
+
 	@Override
 	public String getAuthority() {
 		return getRole();
 	}
-	
-	
+
 }
